Move Validator and Processor to a module
